Move SyncProgressSpec to primitive lib
