+++ lib/primitive/cardano-wallet-primitive.cabal
@@ -175,6 +175,7 @@ test-suite test
     , containers
     , delta-store
     , delta-types
+    , deepseq
     , filepath
     , fmt
     , generic-arbitrary
@@ -203,6 +204,7 @@ test-suite test
   other-modules:
     Cardano.Wallet.Primitive.CollateralSpec
     Cardano.Wallet.Primitive.SlottingSpec
+    Cardano.Wallet.Primitive.SyncProgressSpec
     Cardano.Wallet.Primitive.Types.AddressSpec
     Cardano.Wallet.Primitive.Types.BlockSummarySpec
     Cardano.Wallet.Primitive.Types.CoinSpec
